refactor(pages): Migrate Home page to TypeScript
